Refactor: Bootstrap & Main classes, remove Reflection function
