+++ src/Tickets_Order_Helper.php
@@ -153,9 +153,12 @@ class Tickets_Order_Helper {
 			}
 		} elseif ( ! empty( $this->provider_classname ) ) {
 			$class_reflection   = new ReflectionClass( $this->provider_instance );
-			$attendee_order_key = tribe_call_non_public_method( $this->provider_instance, 'get_attendee_order_key', $class_reflection );
-			$attendee_event_key = tribe_call_non_public_method( $this->provider_instance, 'get_attendee_event_key', $class_reflection );
-			$attendee_object    = tribe_call_non_public_method( $this->provider_instance, 'get_attendee_object', $class_reflection );
+
+			$attendee_helper = new Attendee_Helper();
+
+			$attendee_order_key = $attendee_helper->safely_get_attendee_order_key( $class_reflection );
+			$attendee_event_key = $attendee_helper->safely_get_attendee_event_key( $class_reflection );
+			$attendee_object    = $attendee_helper->safely_get_attendee_object( $class_reflection );
 
 			if (
 				empty( $attendee_order_key )
