--- src/functions/general.php
@@ -1,44 +1,0 @@
-<?php
-// Do not load directly.
-if ( ! defined( 'ABSPATH' ) ) {
-	die( '-1' );
-}
-
-if ( ! function_exists( 'tribe_call_non_public_method' ) ) {
-	/**
-	 * Calls a private/protected method in any class
-	 *
-	 * Only use this as a last resort. Private methods are not intended to
-	 * be accessed, and can change functionality on any update.
-	 *
-	 * To call Tribe__Class::instance()->set_something( $var1, $var2 ) do this:
-	 * tribe_call_non_public_method(
-	 *      Tribe__Class::instance(),
-	 *      'set_something',
-	 *      [ $var1, $var2 ]
-	 * );
-	 *
-	 * @param object $instance The class instance with the private method
-	 * @param string $method   Name of the method
-	 * @param array  $args     Any args to pass to this method
-	 *
-	 * @return mixed|exception Returns method or exception on PHP 5.2
-	 */
-	function tribe_call_non_public_method( $instance, $method, $args = [] ) {
-		if ( version_compare( PHP_VERSION, '5.6', '<' ) ) {
-			$exception = new Exception( sprintf( __( 'The %s function requires PHP version 5.6 or newer to work. Please contact your website host and inquire about updating PHP.', PLUGIN_TEXT_DOMAIN ), __FUNCTION__ ) );
-			_doing_it_wrong( __FUNCTION__, $exception->getMessage(), 'N/A' );
-
-			return $exception;
-		}
-
-		if ( ! is_array( $args ) ) {
-			$args = [ $args ];
-		}
-
-		$reflection_method = new ReflectionMethod( get_class( $instance ), $method );
-		$reflection_method->setAccessible( true );
-
-		return $reflection_method->invokeArgs( $instance, $args );
-	}
-}
